pylons/requirements.txt
-f http://sd2.morpace-i.com/packages/
Beaker==1.5.4
Fabric==0.9.2
FormEncode==1.2.2
Genshi==0.6
Jinja2==2.5.5
Mako==0.3.5
MarkupSafe==0.9.3
MySQL-python==1.2.3
Paste==1.7.4
PasteDeploy==1.3.3
PasteScript==1.7.3
Pygments==1.3.1
Pylons==1.0
Routes==1.12.3
SQLAlchemy==0.6.1
Sphinx==1.0.4
Tempita==0.4
WebError==0.10.2
WebHelpers==1.2
WebOb==0.9.8
WebTest==1.2.1
decorator==3.2.0
docutils==0.7
ipaddr==2.1.5
morpylons==0.0.12
nose==0.11.4
paramiko==1.7.6
pep8==0.6.1
pycrypto==2.0.1
python-ldap==2.3.12
quipp==0.1a6
repoze.what==1.0.9
repoze.what-pylons==1.0
repoze.what.plugins.sql==1.0
repoze.who==1.0.18
repoze.who-friendlyform==1.0.6
repoze.who-testutil==1.0.1
repoze.who.plugins.ldap==1.1a1
repoze.who.plugins.sa==1.0rc2
simplejson==2.1.1
sqlalchemy-migrate==0.6
wsgiref==0.1.2
zope.interface==3.6.1